Move billing preview types to dedicated file
Relocated billingPreviewSchema and related types from zuoraSchemas.ts to a new billingPreview.ts file for better organization and separation of concerns.
